var-install-packages: Restrict var-uuu-installer to i.MX8 and i.MX9 only.
Update the var-release-package so that UUU files are added only for i.MX8/i.MX9.

This avoids build failures on platforms where UUU is not supported.

Also restrict the *.bmap file installation for only i.MX8/i.MX9.

+++ recipes-variscite/var-install-packages/var-release-package.bb
@@ -7,8 +7,11 @@ LIC_FILES_CHKSUM = "file://${COMMON_LICENSE_DIR}/MIT;md5=0835ade698e0bcf8506ecda
 DEPENDS = "\
     tar-native \
     zstd-native \
-    var-uuu-installer \
 "
+
+VAR_UUU_INSTALL = "false"
+VAR_UUU_INSTALL:mx8-nxp-bsp = "true"
+VAR_UUU_INSTALL:mx9-nxp-bsp = "true"
 
 # Default value if the variable is not set
 VAR_RECOVERY_SD_NAME ?= "var-recovery-image-${MACHINE}"
@@ -23,24 +26,30 @@ do_install[depends] += " \
     var-image-swupdate:do_image_complete \
     var-image-swu:do_swuimage \
     var-recovery-image:do_image_complete \
-    var-uuu-installer:do_install \
+    ${@bb.utils.contains('VAR_UUU_INSTALL', 'true', ' var-uuu-installer:do_install', '', d)} \
     var-image-swupdate:do_create_image_sbom_spdx \
     ${VAR_RECOVERY_TARGET_ROOTFS}:do_create_image_sbom_spdx \
 "
 
 do_install () {
-    install -Dm 0644 ${DEPLOY_DIR_IMAGE}/var-uuu-installer.tar.zst ${S}/var-uuu-installer.tar.zst
-    install -Dm 0644 ${DEPLOY_DIR_IMAGE}/var-image-swupdate-${MACHINE}.rootfs.wic.bmap ${S}/${PN}/yocto/var-image-swupdate-${MACHINE}.rootfs.wic.bmap
+    if ${VAR_UUU_INSTALL}; then
+        install -Dm 0644 ${DEPLOY_DIR_IMAGE}/var-image-swupdate-${MACHINE}.rootfs.wic.bmap ${S}/${PN}/yocto/var-image-swupdate-${MACHINE}.rootfs.wic.bmap
+    fi
     install -Dm 0644 ${DEPLOY_DIR_IMAGE}/var-image-swupdate-${MACHINE}.rootfs.wic.zst ${S}/${PN}/yocto/var-image-swupdate-${MACHINE}.rootfs.wic.zst
     install -Dm 0644 ${DEPLOY_DIR_IMAGE}/var-image-swupdate-${MACHINE}.rootfs.spdx.json ${S}/${PN}/yocto/var-image-swupdate-${MACHINE}.rootfs.spdx.json
     install -Dm 0644 ${DEPLOY_DIR_IMAGE}/var-image-swupdate-${MACHINE}.rootfs.json ${S}/${PN}/yocto/var-image-swupdate-${MACHINE}.json
     install -Dm 0644 ${DEPLOY_DIR_IMAGE}/var-image-swu-${MACHINE}.rootfs.swu ${S}/${PN}/yocto/var-image-swu-${MACHINE}.rootfs.swu
-    install -Dm 0644 ${DEPLOY_DIR_IMAGE}/${VAR_RECOVERY_SD_NAME}.wic.bmap ${S}/${PN}/${VAR_RECOVERY_SD_NAME}.wic.bmap
+    if ${VAR_UUU_INSTALL}; then
+        install -Dm 0644 ${DEPLOY_DIR_IMAGE}/${VAR_RECOVERY_SD_NAME}.wic.bmap ${S}/${PN}/${VAR_RECOVERY_SD_NAME}.wic.bmap
+    fi
     install -Dm 0644 ${DEPLOY_DIR_IMAGE}/${VAR_RECOVERY_SD_NAME}.wic.zst ${S}/${PN}/${VAR_RECOVERY_SD_NAME}.wic.zst
     install -Dm 0644 ${DEPLOY_DIR_IMAGE}/${VAR_RECOVERY_TARGET_ROOTFS}-${MACHINE}.rootfs.spdx.json ${S}/${PN}/yocto/${VAR_RECOVERY_TARGET_ROOTFS}-${MACHINE}.rootfs.spdx.json
     install -Dm 0644 ${DEPLOY_DIR_IMAGE}/${VAR_RECOVERY_TARGET_ROOTFS}-${MACHINE}.rootfs.json ${S}/${PN}/yocto/${VAR_RECOVERY_TARGET_ROOTFS}-${MACHINE}.rootfs.json
 
-    tar --zstd -xf var-uuu-installer.tar.zst  -C ${PN}/yocto/
+    if [ -f ${DEPLOY_DIR_IMAGE}/var-uuu-installer.tar.zst ]; then
+        install -Dm 0644 ${DEPLOY_DIR_IMAGE}/var-uuu-installer.tar.zst ${S}/var-uuu-installer.tar.zst
+        tar --zstd -xf var-uuu-installer.tar.zst -C ${PN}/yocto/
+    fi
 
     tar -cf - ${PN} | zstd -19 -T0 -f -o ${PACKAGE_RELEASE_NAME}.tar.zst
